+++ R/norm_data.R
@@ -5,12 +5,11 @@
 #' @details Normalise a dataframe and check for columns correspondance
 #' to be able to use it as an input to create a Ped object.
 #' Multiple test are done and errors are checked.
-#' Sex is calculated based on the `gender` column.
 #'
 #' The `steril` column need to be a boolean either TRUE, FALSE or 'NA'.
 #' Will be considered available any individual with no 'NA' values in the
 #' `available` column.
-#' Duplicated `indId` will nullify the relationship of the individual.
+#' Duplicated `id` will nullify the relationship of the individual.
 #' All individuals with errors will be remove from the dataframe and will
 #' be transfered to the error dataframe.
 #'
@@ -36,23 +35,22 @@
 #' @param ped_df A data.frame with the individuals informations.
 #' The minimum columns required are:
 #'
-#' - `indID` individual identifiers -> `id`
-#' - `fatherId` biological fathers identifiers -> `dadid`
-#' - `motherId` biological mothers identifiers -> `momdid`
-#' - `gender` sex of the individual -> `sex`
-#' - `family` family identifiers -> `famid`
+#' - `id` individual identifiers
+#' - `dadid` biological fathers identifiers
+#' - `momid` biological mothers identifiers
+#' - `sex` of the individual
+#' - `famid` family identifiers
 #'
-#' The `family` column, if provided, will be merged to
+#' The `famid` column, if provided, will be merged to
 #' the *ids* field separated by an underscore using the
 #' [upd_famid()] function.
 #'
 #' The following columns are also recognize and will be transformed with the
 #' [vect_to_binary()] function:
 #'
-#' - `sterilisation` status -> `steril`
-#' - `available` status -> `avail`
-#' - `vitalStatus`, is the individual dead -> `status`
-#' - `affection` status -> `affected`
+#' - `steril` status -> is the individual steril
+#' - `avail` status -> is the individual available
+#' - `deceased` status -> is the individual dead
 #'
 #' The values recognized for those columns are `1` or `0`,
 #' `TRUE` or `FALSE`.
@@ -67,15 +65,14 @@
 #' @include utils.R
 #' @examples
 #' df <- data.frame(
-#'     indId = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
-#'     fatherId = c("A", 0, 1, 3, 0, 4, 1, 0, 6, 6),
-#'     motherId = c(0, 0, 2, 2, 0, 5, 2, 0, 8, 8),
-#'     gender = c(1, 2, "m", "man", "f", "male", "m", "m", "f", "f"),
-#'     available = c("A", "1", 0, NA, 1, 0, 1, 0, 1, 0),
+#'     id = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
+#'     dadid = c("A", 0, 1, 3, 0, 4, 1, 0, 6, 6),
+#'     momid = c(0, 0, 2, 2, 0, 5, 2, 0, 8, 8),
+#'     sex = c(1, 2, "m", "man", "f", "male", "m", "m", "f", "f"),
+#'     avail = c("A", "1", 0, NA, 1, 0, 1, 0, 1, 0),
 #'     famid = c(1, 1, 1, 1, 1, 1, 1, 2, 2, 2),
-#'     sterilisation = c("TRUE", "FALSE", TRUE, FALSE, 1, 0, 1, 0, 1, "TRUE"),
-#'     vitalStatus = c("TRUE", "FALSE", TRUE, FALSE, 1, 0, 1, 0, 1, 0),
-#'     affection = c("TRUE", "FALSE", TRUE, FALSE, 1, 0, 1, 0, 1, 0)
+#'     steril = c("TRUE", "FALSE", TRUE, FALSE, 1, 0, 1, 0, 1, "TRUE"),
+#'     deceased = c("TRUE", "FALSE", TRUE, FALSE, 1, 0, 1, 0, 1, 0)
 #' )
 #' tryCatch(
 #'     norm_ped(df),
@@ -100,20 +97,16 @@ norm_ped <- function(
     )
     err <- data.frame(matrix(NA, nrow = nrow(ped_df), ncol = length(err_cols)))
     colnames(err) <- err_cols
-    cols_need <- c("indId", "fatherId", "motherId", "gender")
-    cols_used <- c(
-        "sex", "steril", "status", "avail", "id", "dadid", "momid", "famid",
-        "error", "affected"
-    )
-    cols_to_use <- c(
-        "available", "family", "sterilisation", "vitalStatus", "affection"
-    )
+    cols_need <- c("id", "dadid", "momid", "sex")
+    cols_used <- c("error")
+    cols_to_use <- c("avail", "famid", "steril", "deceased")
     ped_df <- check_columns(
         ped_df, cols_need, cols_used, cols_to_use, others_cols = TRUE,
         cols_to_use_init = TRUE, cols_used_init = TRUE,
         cols_used_del = cols_used_del
     )
-    ped_df$family[is.na(ped_df$family)] <- missid
+
+    ped_df$famid[is.na(ped_df$famid)] <- missid
 
     if (nrow(ped_df) > 0) {
         ped_df <- dplyr::mutate_if(
@@ -121,26 +114,25 @@ norm_ped <- function(
         )
 
         #### Id #### Check id type
-        for (id in c("indId", "fatherId", "motherId")) {
+        for (id in c("id", "dadid", "momid")) {
             ped_df[[id]] <- as.character(ped_df[[id]])
         }
         err$idErr <- lapply(
             as.data.frame(t(ped_df[, c(
-                "indId", "fatherId", "motherId", "family"
+                "id", "dadid", "momid", "famid"
             )])),
             function(x) {
                 if (any(x == "" & !is.na(x))) {
-                    "One id is Empty"
+                    "one_id_is_empty"
                 } else {
                     NA_character_
                 }
             }
         )
         ## Make a new id from the family and subject pair
-        ped_df$famid <- ped_df$family
-        ped_df$id <- upd_famid(ped_df$indId, ped_df$famid, missid)
-        ped_df$dadid <- upd_famid(ped_df$fatherId, ped_df$famid, missid)
-        ped_df$momid <- upd_famid(ped_df$motherId, ped_df$famid, missid)
+        ped_df$id <- upd_famid(ped_df$id, ped_df$famid, missid)
+        ped_df$dadid <- upd_famid(ped_df$dadid, ped_df$famid, missid)
+        ped_df$momid <- upd_famid(ped_df$momid, ped_df$famid, missid)
 
         ## Set all missid to NA
         ped_df <- dplyr::mutate_at(ped_df, c("id", "dadid", "momid", "famid"),
@@ -148,7 +140,7 @@ norm_ped <- function(
         )
 
         #### Sex ####
-        ped_df$sex <- sex_to_factor(ped_df$gender)
+        ped_df$sex <- sex_to_factor(ped_df$sex)
 
         is_father <- ped_df$id %in% ped_df$dadid & !is.na(ped_df$id)
         is_mother <- ped_df$id %in% ped_df$momid & !is.na(ped_df$id)
@@ -158,9 +150,9 @@ norm_ped <- function(
         ped_df$sex[is_mother] <- "female"
 
         ## Add terminated for sterilized individuals that is neither dad nor mom
-        if ("sterilisation" %in% colnames(ped_df)) {
+        if ("steril" %in% colnames(ped_df)) {
             ped_df$steril <- vect_to_binary(
-                ped_df$sterilisation, logical = TRUE
+                ped_df$steril, logical = TRUE
             )
             ped_df$sex[
                 ped_df$steril == 1 & !is.na(ped_df$steril) &
@@ -170,7 +162,7 @@ norm_ped <- function(
                 (ped_df$sex == "terminated" |
                         ped_df$steril == 1 & !is.na(ped_df$steril)
                 )  & (is_father | is_mother)
-            ] <- "isSterilButIsParent"
+            ] <- "is_steril_but_is_parent"
             ped_df$steril[!is.na(err$sexErrTer) &
                     (is_father | is_mother)
             ] <- FALSE
@@ -179,14 +171,14 @@ norm_ped <- function(
         ## Check error between sex and parentality
         err$sexNA[!ped_df$sex %in%
                 c("male", "female", "terminated", "unknown")
-        ] <- "sexNotRecognise"
-        err$sexErrMoFa[is_father & is_mother] <- "isMotherAndFather"
+        ] <- "sex_not_recognise"
+        err$sexErrMoFa[is_father & is_mother] <- "is_mother_and_father"
         err$sexErrFa[
             is_father & ped_df$sex != "male"
-        ] <- "isFatherButNotMale"
-        err$sexErrMo[is_mother &
-                ped_df$sex != "female"
-        ] <- "isMotherButNotFemale"
+        ] <- "is_father_but_not_male"
+        err$sexErrMo[
+            is_mother & ped_df$sex != "female"
+        ] <- "is_mother_but_not_female"
 
         ## Unite all sex errors in one column
         err <- tidyr::unite(
@@ -209,18 +201,18 @@ norm_ped <- function(
         ## Register errors
         err$idErrFa[ped_df$dadid %in% id_duplicated &
                 !is.na(ped_df$dadid)
-        ] <- "fatherIdDuplicated"
+        ] <- "dadid_duplicated"
         err$idErrMo[ped_df$momid %in% id_duplicated &
                 !is.na(ped_df$momid)
-        ] <- "motherIdDuplicated"
+        ] <- "momid_duplicated"
         err$idErrSelf[ped_df$id %in% id_duplicated &
                 !is.na(ped_df$id)
-        ] <- "selfIdDuplicated"
-        err$idErrOwnParent[ped_df$id %in% id_own_parent] <- "isItsOwnParent"
+        ] <- "self_id_duplicated"
+        err$idErrOwnParent[ped_df$id %in% id_own_parent] <- "is_its_own_parent"
         err$idErrBothParent[
-            (ped_df$dadid %in% missid & (!ped_df$momid %in% missid)) |
-                ((!ped_df$dadid %in% missid) & ped_df$momid %in% missid)
-        ] <- "oneParentMissing"
+            (ped_df$dadid %in% missid & (!ped_df$momid %in% missid))
+            | ((!ped_df$dadid %in% missid) & ped_df$momid %in% missid)
+        ] <- "one_parent_missing"
 
         ## Unite all id errors in one column
         err <- tidyr::unite(
@@ -232,16 +224,12 @@ norm_ped <- function(
         err$idError[err$idError == ""] <- NA
 
         #### Available ####
-        if ("available" %in% colnames(ped_df)) {
-            ped_df$avail <- vect_to_binary(ped_df$available, logical = TRUE)
+        if ("avail" %in% colnames(ped_df)) {
+            ped_df$avail <- vect_to_binary(ped_df$avail, logical = TRUE)
         }
-        #### Status ####
-        if ("vitalStatus" %in% colnames(ped_df)) {
-            ped_df$status <- vect_to_binary(ped_df$vitalStatus, logical = TRUE)
-        }
-        #### Affected ####
-        if ("affection" %in% colnames(ped_df)) {
-            ped_df$affected <- vect_to_binary(ped_df$affection, logical = TRUE)
+        #### Deceased ####
+        if ("deceased" %in% colnames(ped_df)) {
+            ped_df$deceased <- vect_to_binary(ped_df$deceased, logical = TRUE)
         }
 
         #### Convert to num ####
@@ -337,7 +325,7 @@ norm_rel <- function(rel_df, na_strings = c("NA", ""), missid = NA_character_) {
         rel_df$code <- rel_code_to_factor(rel_df$code)
         err$codeErr[!rel_df$code %in%
                 c("MZ twin", "DZ twin", "UZ twin", "Spouse")
-        ] <- "CodeNotRecognise"
+        ] <- "code_not_recognise"
 
         #### Check for id errors #### Set ids as characters
         rel_df <- rel_df %>%
@@ -346,8 +334,8 @@ norm_rel <- function(rel_df, na_strings = c("NA", ""), missid = NA_character_) {
         ## Check for non null ids
         len1 <- nchar(rel_df$id1)
         len2 <- nchar(rel_df$id2)
-        err$id1Err[is.na(len1) | len1 %in% missid] <- "indId1length0"
-        err$id2Err[is.na(len2) | len2 %in% missid] <- "indId2length0"
+        err$id1Err[is.na(len1) | len1 %in% missid] <- "id1_length0"
+        err$id2Err[is.na(len2) | len2 %in% missid] <- "id2_length0"
 
         ## Compute id with family id
         rel_df$id1 <- upd_famid(rel_df$id1, rel_df$famid, missid)
@@ -357,7 +345,7 @@ norm_rel <- function(rel_df, na_strings = c("NA", ""), missid = NA_character_) {
             ~replace(., . %in% c(na_strings, missid), NA_character_)
         )
 
-        err$sameIdErr[rel_df$id1 == rel_df$id2] <- "SameId"
+        err$sameIdErr[rel_df$id1 == rel_df$id2] <- "same_id"
 
         ## Unite all id errors in one column
         rel_df$error <- unite(err, "error",